refactor(desktop): extract web search ipc
