+++ apps/desktop/src/main/__tests__/main-process-contract-source-helpers.ts
@@ -21,6 +21,7 @@ export const MAIN_PROCESS_SOURCE_REPO_PATHS: readonly string[] = [
   'apps/desktop/src/main/subscription-ipc-main.ts',
   'apps/desktop/src/main/system-prompt-main.ts',
   'apps/desktop/src/main/usage-ipc-main.ts',
+  'apps/desktop/src/main/web-search-ipc-main.ts',
   'apps/desktop/src/main/workspace-resources-ipc-main.ts',
 ];
