+++ apps/desktop/src/main/main.ts
@@ -32,15 +32,8 @@ import type {
   UpdateAppSettingsResult,
   UpdateAppSettingsInput,
 } from '@maka/core';
-import {
-  isWebSearchProvider,
-  normalizeWebSearchLimit,
-  normalizeWebSearchQuery,
-} from '@maka/core';
-import { queryTavily, TAVILY_TEST_QUERY, TAVILY_TEST_LIMIT } from './web-search/tavily.js';
 import { buildWebSearchAgentTool, WEB_SEARCH_TOOL_NAME } from './web-search/agent-tool.js';
 import { buildRiveWorkflowTool } from './rive-workflow-tool.js';
-import { resolveTavilyApiKey } from './web-search/credentials.js';
 import { runThreadSearch } from './search/thread-search.js';
 import {
   persistArchivedToolResultToArtifacts,
@@ -181,6 +174,7 @@ import { registerPlanReminderIpc } from './plan-reminders-ipc-main.js';
 import { registerWorkspaceResourcesIpc } from './workspace-resources-ipc-main.js';
 import { registerDailyReviewIpc } from './daily-review-ipc-main.js';
 import { registerUsageIpc } from './usage-ipc-main.js';
+import { registerWebSearchIpc } from './web-search-ipc-main.js';
 
 const buildInfo = resolveBuildInfo(app.isPackaged, app.getAppPath());
 
@@ -1041,67 +1035,7 @@ function registerIpc(): void {
     emitConnectionListChanged,
   });
 
-  // PR-WEB-SEARCH-TAVILY-0: explicit user-triggered web search. Token
-  // is read from settings inside main; renderer never sees it. Falls
-  // back to the `apiKey` carried by the request only when present (the
-  // Settings "测试" button passes a draft key so the user can validate
-  // before saving). Incognito workspaces fail closed before fetch.
-  const unsupportedWebSearchProviderResponse = {
-    ok: false,
-    reason: 'unsupported_provider' as const,
-    message: '当前配置不支持这个搜索引擎，请选择 Tavily 后重试。',
-  };
-  ipcMain.handle(
-    'web-search:query',
-    async (
-      _event,
-      request: { query?: unknown; limit?: unknown; provider?: unknown; apiKey?: unknown },
-    ) => {
-      const provider = request?.provider;
-      if (provider !== undefined && !isWebSearchProvider(provider)) {
-        return unsupportedWebSearchProviderResponse;
-      }
-      const query = normalizeWebSearchQuery(request?.query);
-      if (query === null) {
-        return { ok: false, reason: 'invalid_query' as const, message: '请输入有效的搜索关键词。' };
-      }
-      const privacy = await getWorkspacePrivacyContext();
-      if (privacy.incognitoActive) {
-        return { ok: false, reason: 'incognito_active' as const, message: '隐身模式下禁用联网搜索。' };
-      }
-      const settings = await settingsStore.get();
-      if (!settings.webSearch.enabled) {
-        return {
-          ok: false,
-          reason: 'not_configured' as const,
-          message: '请先在 设置 · 联网搜索 中启用 Tavily。',
-        };
-      }
-      const effectiveKey = resolveTavilyApiKey({ settings, draftKey: request?.apiKey });
-      const limit = normalizeWebSearchLimit(request?.limit);
-      return queryTavily({ apiKey: effectiveKey, query, limit });
-    },
-  );
-
-  ipcMain.handle(
-    'web-search:test',
-    async (
-      _event,
-      request: { provider?: unknown; apiKey?: unknown } | undefined,
-    ) => {
-      const provider = request?.provider;
-      if (provider !== undefined && !isWebSearchProvider(provider)) {
-        return unsupportedWebSearchProviderResponse;
-      }
-      const settings = await settingsStore.get();
-      const effectiveKey = resolveTavilyApiKey({ settings, draftKey: request?.apiKey });
-      return queryTavily({
-        apiKey: effectiveKey,
-        query: TAVILY_TEST_QUERY,
-        limit: TAVILY_TEST_LIMIT,
-      });
-    },
-  );
+  registerWebSearchIpc({ settingsStore, getWorkspacePrivacyContext });
 
   ipcMain.handle('search:thread', async (_event, request: unknown) => {
     // PR-SEARCH-2 review fixup (@xuan `2f1aba55`): pass `unknown`
